Remove part of README install instructions that are obsolete They refer to bash script names that no longer exist in the vm-ubuntu-20.04 directory.  Updating the names of the scripts is not sufficient to give orking instructions, since the current bash scripts currently hard-code path names of /home/vagrant and/or /home/p4 directory names that will not typically exist on an arbitrary Ubuntu Linux system without extra steps to create them, and they also assume several files exist in particular directories, as they are copied by the Vagrantfile.
The bash scripts could of course be updated to work when run from the
Vagrantfile, and also from a user's freshly installed Ubuntu Linux
system, but I am not going to attempt to do that with this commit.
Someone else is welcome to make such changes if they are interested in
making that work.